requirements.txt
flast
tensorflow
opencv-python
numpy
pillow